+++ src/main.ts
@@ -15,16 +15,44 @@ import 'wikiparser-node/extensions/typings.d.ts';
 import type {Config} from 'wikiparser-node';
 import type * as Monaco from 'monaco-editor';
 import type {LanguageInput} from 'shiki/core';
+import type {IRawGrammar} from '@shikijs/core/textmate';
 import type {IWikitextModel} from './linter.ts';
 
-const wikitext: LanguageInput = require('../vendor/wikitext.tmLanguage.json'),
+declare type IRawRule = IRawGrammar['repository'][string];
+
+const wikitext: LanguageInput & IRawGrammar = require('../vendor/wikitext.tmLanguage.json'),
 	config: Monaco.languages.LanguageConfiguration = require('../vendor/language-configuration.json'),
 	defaultConfig: Config = require('wikiparser-node/config/default.json');
+const repository = wikitext.repository['wikitext']!.repository!,
+	magicWords = repository['magic-words']!.repository!,
+	variables = magicWords['variables']!.patterns!,
+	behaviorSwitches = magicWords['behavior-switches']!.patterns!;
+
+const defineGrammar = (rule: IRawRule, options: string[], key: 'match' | 'begin' = 'match'): void => {
+	Object.assign(rule, {[key]: rule[key]!.replace('$1', options.join('|'))});
+};
 
 const registerWiki = async (monaco: typeof Monaco, parserConfig: Config | boolean = false): Promise<void> => {
-	const tempConfig = typeof parserConfig === 'object' ? parserConfig : defaultConfig;
+	const tempConfig = typeof parserConfig === 'object' ? parserConfig : defaultConfig,
+		{doubleUnderscore, redirection, parserFunction, nsid, protocol, ext} = tempConfig;
+	if (doubleUnderscore[0].length === 0) {
+		doubleUnderscore[0] = Object.keys(doubleUnderscore[2]!);
+	}
+	defineGrammar(repository['redirect']!, redirection);
+	defineGrammar(
+		variables[0]!,
+		[
+			...Object.keys(parserFunction[0]).filter(s => !s.startsWith('#')),
+			...(parserFunction.slice(2) as string[][]).flat(),
+		],
+	);
+	defineGrammar(variables[1]!, parserFunction[1].filter(s => !s.startsWith('#')));
+	defineGrammar(behaviorSwitches[0]!, doubleUnderscore[0]);
+	defineGrammar(behaviorSwitches[1]!, doubleUnderscore[1]);
+	defineGrammar(repository['file-link']!, Object.entries(nsid).filter(([, v]) => v === 6).map(([k]) => k), 'begin');
+	defineGrammar(repository['external-link']!, [protocol.replace(/\//gu, String.raw`\/`), String.raw`\/\/`]);
 	config.autoClosingPairs!.push(
-		...[tempConfig.ext, tempConfig.html.slice(0, 2)].flat(2).map(tag => ({open: `<${tag}>`, close: `</${tag}>`})),
+		...[ext, tempConfig.html.slice(0, 2)].flat(2).map(tag => ({open: `<${tag}>`, close: `</${tag}>`})),
 	);
 	const highlighter = await createHighlighterCore({
 		langs: [